Fix DirectoryNotFoundException on build

+++ build/BuildSteps.cs
@@ -112,10 +112,13 @@ namespace Build
             {
                 // Python worker's dependencies are platform dependent and need to be copied accordingly
                 var pythonDir = Path.Combine(Settings.OutputDir, runtime, "workers", "python");
-                var allOsDirectories = Directory.GetDirectories(pythonDir);
-                var pythonPlatformSpecificWorker = Path.Combine(pythonDir, Settings.RuntimesToOS[runtime]);
-                FileHelpers.RecursiveCopy(pythonPlatformSpecificWorker, pythonDir);
-                allOsDirectories.ToList().ForEach(dir => Directory.Delete(dir, recursive: true));
+                if (Directory.Exists(pythonDir))
+                {
+                    var allOsDirectories = Directory.GetDirectories(pythonDir);
+                    var pythonPlatformSpecificWorker = Path.Combine(pythonDir, Settings.RuntimesToOS[runtime]);
+                    FileHelpers.RecursiveCopy(pythonPlatformSpecificWorker, pythonDir);
+                    allOsDirectories.ToList().ForEach(dir => Directory.Delete(dir, recursive: true));
+                }
             }
         }
